Sonar fixes
