+++ designer/server/src/routes/admin/dead-letter-queue-helper.js
@@ -1,0 +1,109 @@
+import { formAdapterSubmissionMessagePayloadSchema } from '@defra/forms-engine-plugin/engine/types/schema.js'
+
+import { getDeadLetterQueueMessages } from '~/src/lib/dead-letter-queue.js'
+import { createJoiError } from '~/src/lib/error-boom-helper.js'
+
+const MAX_DLQ_READ_ATTEMPTS = 5
+const WAIT_TIME_IN_MILLIS = 1000
+
+/**
+ * @param {string} messageJson
+ */
+export function validateMessageJson(messageJson) {
+  let json
+  try {
+    json = JSON.parse(messageJson)
+  } catch (err) {
+    const typedError = /** @type {{ message?: string }} */ (err)
+    return {
+      error: createJoiError(
+        'messageJson',
+        `Invalid JSON: ${typedError.message}`
+      )
+    }
+  }
+
+  /**
+   * @type { FormAdapterSubmissionMessagePayload | undefined }
+   */
+  const messageBody = json.Body
+  if (!messageBody) {
+    return {
+      error: createJoiError(
+        'messageJson',
+        'Invalid JSON: Missing "Body" element'
+      )
+    }
+  }
+
+  const { error } = formAdapterSubmissionMessagePayloadSchema.validate(
+    messageBody,
+    {
+      abortEarly: false,
+      stripUnknown: true
+    }
+  )
+  if (error) {
+    const errorText = error.details.map((d) => d.message).join(', ')
+    const joiError = createJoiError(
+      'messageJson',
+      `JSON does not match the schema: ${errorText}`
+    )
+    return {
+      error: joiError
+    }
+  }
+  return {
+    body: messageBody
+  }
+}
+
+/**
+ * @param {DeadLetterQueues} dlq
+ * @param {string} token
+ * @param {string} messageId
+ * @returns {Promise<any | undefined>}
+ */
+export async function readDlqWithRetry(dlq, token, messageId) {
+  let attempts = 0
+  let foundMessage
+  while (attempts < MAX_DLQ_READ_ATTEMPTS) {
+    attempts++
+    const messages = await getDeadLetterQueueMessages(dlq, token)
+
+    foundMessage = messages.find((m) => m.MessageId === messageId)
+    if (foundMessage) {
+      break
+    }
+
+    await new Promise((resolve) => setTimeout(resolve, WAIT_TIME_IN_MILLIS))
+  }
+  return foundMessage
+}
+
+/**
+ * Keep specific properties
+ * @param {any} message
+ * @returns {{ json: { MessageId: string, Body: any }}}
+ */
+export function dlqMessageMapper(message) {
+  return {
+    json: {
+      MessageId: message.MessageId,
+      Body: JSON.parse(message.Body)
+    }
+  }
+}
+
+/**
+ * @param {any[]} messages
+ * @returns {{ json: { MessageId: string, Body: any }}[]}
+ */
+export function dlqMessagesMapper(messages) {
+  return messages.map(dlqMessageMapper)
+}
+
+/**
+ * @import { DeadLetterQueues } from '@defra/forms-model'
+ * @import { FormAdapterSubmissionMessagePayload } from '@defra/forms-engine-plugin/engine/types.js'
+ */
